requirements.txt
asgiref==3.11.0
Django==6.0.1
sqlparse==0.5.5
tensorflow>=2.16.0
Pillow>=10.0.0
numpy>=1.26.0
opencv-python>=4.8.0
